Extract fallback finding normalization
